refactor: extract GetRequestedConfigTrait trait from ConfigFactory
Since the main details will remain approximately the same between the
factory and the abstract factory, but the signatures will differ, we can
use a trait to accomplish them.
